Added environment variables so pipenv will create a local venv.
pipenv will use existing environment variables such as WORKON_HOME
to decide where to install the virtualenv. The name given to the
virtualenv uses the parent directory and a hash of the full path.
Since the parent directory is 'backend' if this template is used
for more than one project then we will end up with a set of similarly
named environments which is problematic. To fix this we force pipenv
to create the virtualenv local to the project.